updating my repo

+++ client/src/Components/LoginForm.js
@@ -17,26 +17,17 @@ const LoginForm = () => {
     try {
       const { data } = await axios.post("/users/login", form);
 
-      //   confirmation logic starts here
-      // if (!data.user.confirmed) {
-      //   history.push("/confirm");
-      // } else {
-      //   setUserData({
-      //     token: data.token,
-      //     user: data.user,
-      //   });
+      if (!data.user.confirmed) {
+        history.push("/confirm");
+      } else {
+        setUserData({
+          token: data.token,
+          user: data.user,
+        });
 
-      //   localStorage.setItem("auth-token", data.token);
-      //   history.push("/");
-      // }
-
-      setUserData({
-        token: data.token,
-        user: data.user,
-      });
-
-      localStorage.setItem("auth-token", data.token);
-      history.push("/");
+        localStorage.setItem("auth-token", data.token);
+        history.push("/");
+      }
     } catch (err) {
       console.log("err", err.response);
       toast.error(err.response.data.msg);
